sampling_cost_smoke: three %zu conversions were reading eight bytes for a four-byte argument
num_states() and num_canonical_states() return uint32_t and were passed to printf
under %zu, which is size_t. In a varargs call that reads eight bytes where four were
passed, so every conversion after it in the same call is reading from the wrong
place. It is undefined behaviour, and the reason the output has looked correct is
that the upper half of the register happened to be zero.

Three sites, both printf calls in the file. num_events() was already cast two lines
below one of them, which is what the others should have done.

The build did not catch it because the target is compiled without -Wall; with it,
g++ reports all three as -Wformat. The values are unchanged after the fix --
states=409,114, canonical=1,205, events=409,113 on growth:1:9 at four threads,
the same figures every measurement tonight was taken against.

+++ tools/sampling_cost_smoke.cpp
@@ -309,7 +309,9 @@ int main(int argc, char** argv) {
 
     std::printf("done %.2f ms  states=%zu canon=%zu events=%zu matches=%zu drained=%zu\n",
                 std::chrono::duration<double, std::milli>(t1 - t0).count(),
-                hg.num_states(), hg.num_canonical_states(), hg.num_events(),
+                static_cast<size_t>(hg.num_states()),
+                static_cast<size_t>(hg.num_canonical_states()),
+                static_cast<size_t>(hg.num_events()),
                 e.total_matches(), e.states_drained());
 
     // ONE MACHINE-READABLE LINE CARRYING EVERY COUNT THIS RUN ALREADY COMPUTED. The `done` line
@@ -359,7 +361,8 @@ int main(int argc, char** argv) {
                              : size_t{1}),
                 edges, steps, threads, canon.c_str(),
                 std::chrono::duration<double, std::milli>(t1 - t0).count(),
-                hg.num_states(), hg.num_canonical_states(),
+                static_cast<size_t>(hg.num_states()),
+                static_cast<size_t>(hg.num_canonical_states()),
                 static_cast<size_t>(hg.num_events()), e.total_matches(),
                 cg.num_causal_edges(), cg.num_causal_event_pairs(),
                 cg.num_branchial_edges(), cg.num_branchial_pairs_claimed(),
